\name{LongdoDict}
\alias{LongdoDict}
\title{
Longdo Dictionary (https://dict.longdo.com/)
}
\description{
Search an English or Thai Word from Longdo Dictionary in RStudio
}
\usage{
LongdoDict(word)
}
\arguments{
  \item{word}{
An English or Thai word that you would like to find its meaning from the Longdo Dictionary (https://dict.longdo.com/)
}
}
\references{
https://dict.longdo.com/
}
\author{
Sompob Saralamba <saralamba@gmail.com>
}
\examples{
\code{LongdoDict("freedom")}
}
