\name{LongdoDict}
\alias{LongdoDict}
\title{
Longdo Dictionary (https://dict.longdo.com/)
}
\description{
Lookup English or Thai words using the Longdo Dictionary within RStudio.
}
\usage{
LongdoDict(word)
}
\arguments{
  \item{word}{
An English or Thai word you wish to look up for its meaning on the Longdo Dictionary (https://dict.longdo.com/).
}
}
\references{
https://dict.longdo.com/
}
\author{
Sompob Saralamba <saralamba@gmail.com>
}
\examples{
\code{LongdoDict("freedom")}
}
